Add modifications for assigment 4
b. Modify your finder-test.sh to run with necessary files found in the PATH.
i. In other words, you should be able to run /path/to/script/finder-test.sh and the script should run successfully,
assuming all executables are in the PATH and config files are at /etc/finder-app/conf. This change will be added to your assignment 3 and later source repository.
The point of this step is to ensure you can run the finder-test.sh script using /usr/bin/finder-test.sh and all scripts and executables needed by finder-test.sh will be located on the target qemu rootfs.
c. Modify your finder-test.sh script to write a file with output of the finder command to /tmp/assignment4-result.txt